+++ src/NCalc.Core/Factories/IExpressionFactory.cs
@@ -4,9 +4,13 @@ public interface IExpressionFactory
 {
     public Expression Create(
         string expression,
-        ExpressionConfiguration? configuration = null);
+        ExpressionConfiguration? configuration = null,
+        ExpressionContext? context = null,
+        CultureInfo? cultureInfo = null);
 
     public Expression Create(
         LogicalExpression logicalExpression,
-        ExpressionConfiguration? configuration = null);
+        ExpressionConfiguration? configuration = null,
+        ExpressionContext? context = null,
+        CultureInfo? cultureInfo = null);
 }